feat(editor): Add support for latex features
This is experimental. Vimtex has lots of features, but right now we care
more the autocompile feature. Enabling the LSP stuff, makes the
experience comparable to, if not better than, overleaf. The only thing
is missing is to directly edit overleaf pages (there is an extension,
but it needs to be validated).